+++ packages/react/src/components/display/Divider/Divider.test.tsx
@@ -1,9 +1,10 @@
-import { vars } from '@blockle/blocks-core';
+import { getNameFromVanillaCSSVar, vars } from '@blockle/blocks-core';
 import { screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
 import { render } from '../../../testUtils/testUtils.js';
 import { Divider } from './Divider.js';
+import * as styles from './divider.css.js';
 
 describe('Divider', () => {
   it('should render with storybook', async () => {
@@ -43,13 +44,15 @@ describe('Divider', () => {
   });
 
   // TODO: Fix this test
-  it.skip('should apply color style', () => {
+  it('should apply color style', () => {
     render(<Divider color="primary-500" />);
 
     const divider = screen.getByRole('separator');
-    expect(getComputedStyle(divider).getPropertyValue('--test-var')).toEqual(
-      vars.color['primary-500'],
-    );
+    expect(
+      getComputedStyle(divider).getPropertyValue(
+        getNameFromVanillaCSSVar(styles.dividerColorVar),
+      ),
+    ).toEqual(vars.color['primary-500']);
   });
 
   it('should have horizontal orientation', () => {