+++ packages/react/src/components/form/Button/Button.tsx
@@ -29,7 +29,6 @@ export type ButtonProps = {
   intent?: ButtonTheme['variants']['intent'];
   loading?: boolean;
   popovertarget?: string;
-  ref?: React.Ref<HTMLButtonElement>;
   size?: ButtonTheme['variants']['size'];
   startSlot?: React.ReactNode;
   type?: 'button' | 'submit' | 'reset';
@@ -47,7 +46,6 @@ export const Button: React.FC<ButtonProps> = ({
   endSlot,
   intent,
   loading,
-  ref,
   size,
   startSlot,
   type,
@@ -77,7 +75,6 @@ export const Button: React.FC<ButtonProps> = ({
 
   return (
     <Template
-      ref={ref}
       asChild={asChild}
       disabled={disabled || loading}
       className={classnames(
@@ -86,7 +83,6 @@ export const Button: React.FC<ButtonProps> = ({
         atoms(atomsProps),
         className,
       )}
-      // Do not pass type attribute if using asChild since child component may not support it
       type={asChild ? undefined : (type ?? 'button')}
       {...otherProps}
     >